+++ src/editor/error-reason.ts
@@ -2,6 +2,7 @@ export { ErrorReason, toString };
 
 enum ErrorReason {
   DidNotFindTypeToExtract,
+  DidNotFindMultipleDeclarationsToSplit,
   CantImportReferences,
   DidNotFindOtherFiles,
   DidNotFindCodeToMove,
@@ -54,6 +55,9 @@ function toString(reason: ErrorReason): string {
   switch (reason) {
     case ErrorReason.DidNotFindTypeToExtract:
       return didNotFind("a type to extract");
+
+    case ErrorReason.DidNotFindMultipleDeclarationsToSplit:
+      return didNotFind("multiple variable declarations to split");
 
     case ErrorReason.CantImportReferences:
       return cantDoIt(