+++ src/refactorings/split-multiple-declarations/split-multiple-declarations.ts
@@ -1,0 +1,58 @@
+import { Editor, ErrorReason } from "../../editor/editor";
+import { Selection } from "../../editor/selection";
+import * as t from "../../ast";
+
+export {
+  splitMultipleDeclarations,
+  createVisitor as canSplitMultipleDeclarations
+};
+
+async function splitMultipleDeclarations(editor: Editor) {
+  const { code, selection } = editor;
+  const updatedCode = updateCode(t.parse(code), selection);
+
+  if (!updatedCode.hasCodeChanged) {
+    editor.showError(ErrorReason.DidNotFindMultipleDeclarationsToSplit);
+    return;
+  }
+
+  await editor.write(updatedCode.code);
+}
+
+function updateCode(ast: t.AST, selection: Selection): t.Transformed {
+  return t.transformAST(
+    ast,
+    createVisitor(selection, (path: t.NodePath<t.VariableDeclaration>) => {
+      const declarators = path.node.declarations;
+      const kind = path.node.kind;
+
+      const declarations = declarators.map(function (declarator) {
+        return t.variableDeclaration(kind, [declarator]);
+      });
+
+      path.replaceWithMultiple(declarations);
+    })
+  );
+}
+
+function createVisitor(
+  selection: Selection,
+  onMatch: (path: t.NodePath<t.VariableDeclaration>) => void
+): t.Visitor {
+  return {
+    VariableDeclaration(path) {
+      selection;
+
+      const declarations = path.node.declarations;
+      if (!hasInitializedMultipleDeclarations(declarations)) return;
+
+      onMatch(path);
+    }
+  };
+}
+
+function hasInitializedMultipleDeclarations(
+  declarations: t.VariableDeclarator[]
+): boolean {
+  return declarations.length >= 2;
+}