+++ CHANGELOG.md
@@ -14,6 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - **[New Refactoring]** Extract Type! It works exactly like Extract Variable, but for types. Even more convenient: the shortcut is the same (`Ctrl + Alt + V` / `⌥ ⌘ V` on MacOS). If you're on a _variable_, it will extract the _variable_. If you're on a _type_, it will extract the _type_ 😉
+- **[New Refactoring]** Split Multiple Declarations, to create distinct variable declarations from a single one. Thanks to @iulspop for this one!
+
+### Fixed
+
+- Split Declaration and Initialization would duplicate type annotations in TypeScript. @iulspop fixed that 🔥
 
 ## [4.13.1] - 2021-01-10